Remove Renpho from docs (#373)
It appears that Renpho is now restricting their old app that we support:
>We apologize for the inconvenience, but the Renpho App is currently
unable to register data. To continue using our services, we kindly
request that you switch to the Renpho Health App [...]

https://apps.apple.com/us/app/renpho-outdated-version/id1219889310

I just had a user try, and it won't let them sign up for an account.

Brought this up here:
https://junctionapi.slack.com/archives/C074VMJ7HRR/p1752782441899069
Maitham confirmed below the thread that it should be removed.